Added option to fix some of the b values, and example for how to compare uncertainty measures vs some other choice of b values
Hi Eric,
 
Yes, both of those extensions/calculations should be easy.
 
How about I add them, and then we discuss?
 
Diego
 
 
 
From: Sigmund, Eric <Eric.Sigmund@nyulangone.org>
Date: Thursday, October 24, 2024 at 2:52 PM
To: DIEGO HERNANDO <dhernando@wisc.edu>
Subject: RE: [Extern] [External] IVIM white paper section brainstorming

Hi Diego,
 
Hope you are well.  I wondered if I could pick your brain on another slight variant of the CLRB algorithm that we are using to suggest some b-value ranges in the white paper.  We have come around to using the ranges functionality you added as well as an upper limit on maximum b-value which varies for each organ (driven by usage patterns and SNR effects due to variable T2).  Since we are also including a defined ‘threshold’ value to support segmented fitting, which we determined could be fulfilled (that is achieve suppression of perfusion compartment by exp(-3)) by 200 s/mm2 for all cases except brain which would need 300 s/mm2.  As of now I have been ‘adding’ this value to those determined by CLRB to achieve final lists, which may not be ideal and could in some cases be redundant.  Is it possible to start with certain b-values ab initio and then have the algorithm add them?
 
This also relates to another question I have; the kidney values we determine by this standard process are similar but not identical to the ones we suggested in a prior consensus paper.  I wondered if it is possible to compare the precision expected from the CLRB selected values with a specified set to see how much benefit would be afforded and whether it is worth altering the suggestion.
 
Again thanks again for all your efforts in this area.  We have been editing the full manuscript aggressively and hope to circulate a full version to the group soon.
 
Thanks,
 
Eric